requirements.txt
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-dateutil>=2.8.0
flask>=2.0.0
